Add ability to go back to last PR
You can now respond `back` when prompted for input in the main loop to
see the last PR in case you made a mistake. Also pulled in the
`pull-requests` script to check for missing PRs. Wired it up to run with
`npm run pulls < repo >`. Did some refactoring so that the two scripts
can share some code.