Increase threadpool size for all three scripts
Today, we encountered an error where the HTTPS request would fail with
the message `ETIMEDOUT`. After some googling, it looks like this happens
because by default node only allocates 4 threads to DNS resolution. So,
if you make a bunch of HTTP requests in a loop (which this script does),
those DNS requests will start blocking each other and some will time
out. This commit adds code to increase the threadpool size for the node
process to the maximum at the start of each script. Might fix it?